# openapi.yaml
openapi: 3.1.0
info:
  title: REU.ng API
  version: 2.0.0
  description: >
    Enterprise-grade OpenAPI 3.1 specification for the Growth Estate / Reu.ng application.
    Covers authentication (JWT), email verification, password reset, KYC, lands, purchases,
    deposits (Paystack · Monnify · OPay), withdrawals, webhooks, bank resolution,
    notifications, transaction PIN, referrals, portfolio, support tickets, and admin routes.
servers:
  - url: https://growth-estate.onrender.com/api
    description: Production
  - url: http://127.0.0.1:8000/api
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    userId:
      name: user
      in: path
      required: true
      schema:
        type: integer
      description: User ID
...
paths:
...
  /admin/users/{user}:
    get:
      tags: [Admin – Users]
      summary: Get a single user with recent transactions and KYC
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        "200":
          description: User detail
    delete:
      tags: [Admin – Users]
      summary: Delete a non-admin user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        "200":
          description: User deleted
        "403":
          description: Cannot delete admin or self
...